Remove a pass of comparing vectors for push function

+++ projects/liars_poker_bot/src/collections/bitarray.rs
@@ -1,6 +1,6 @@
 #[derive(Default, Clone, Copy, PartialEq, Hash, Eq)]
 pub struct BitArray {
-    pub values: u128,
+    pub values: u32,
 }
 
 impl BitArray {
